Fix: Java agent: update ref count in enabledLoggers
Integer objects are immutable in Java, so

    Integer refcount = enabledLoggers.get(name);
    refcount--;

does not update the value in enabledLoggers.

However, this bug "fixes" a bug in the session daemon
(LTTng-tools) which sends multiple _disable event_
commands for the same event name (one when the event is
disabled manually, and another one when the session is
destroyed). Therefore we use the registration done
command's version to know if the Java agent is connected to
a fixed or non-fixed session daemon, using the old
behaviour if connected to a non-fixed one.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>