Rename helper providers and events for consistency
This patch renders uniform the provider names of UST helpers by making
them all start with the "lttng_ust_" prefix. Also, the helper formerly
known as "ust_baddr" is now "lttng_ust_dl", and its "push" and "pop"
events are now "dlopen" and "dlclose", respectively. This is in line
with the other helpers, for which the name of the provider reflects
the name of the library, and the names of events correspond to those
of the traced functions.

Signed-off-by: Antoine Busque <abusque@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>