Fix: handle sys_futex EINTR and EWOULDBLOCK
Handling sys_futex EINTR allows us to retry waiting on the futex
immediately if we are interrupted by a signal without having to try
connecting to the session daemon.

This should not impact correctness though, as the wait scheme would
simply try to connect, fail, and try waiting again.

Also handle errors on FUTEX_WAKE.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>