Small refactor of the Java agent's TCP client
Better separate the "commands" (sent from the sessiond to the
agent) from the "responses" (sent from the agent to the sessiond
as replies to commands) into distinct objects.

ISessiondCommand#execute() is now an interface method, and it
returns an ILttngAgentResponse.

This more rigorous handling of commands and responses will make
it easier to add support for additional commands in the future.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>