Add ctf_string_maybe_null() field macro
This one has the same signature as ctf_string(), but it will
write "(null)", like printf() does with a NULL %s, instead of
crashing because ctf_string() calls strlen() on its argument.

This is implemented as a different macro because otherwise, using
ctf_string(), you would not know if the argument is NULL or if
the argument was really "(null)". By using ctf_string_maybe_null(),
your code indicates that you know that the string might be NULL.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>