Restrict Java context retriever names to a set of valid characters
Since the context/retriever names end up as-is as part of the
metadata, only alphanumerical characters, periods "." and
underscores "_" should be accepted. The names must also not
start with a number.

Refuse registration of retriever names that do not respect these
conditions, so that the problem is reported right away to the
application.

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>