Introduce lttng_ust_loaded weak symbol
This variable can be tested by applications to check whether lttng-ust
is loaded. They simply have to define their own "lttng_ust_loaded" weak
symbol, and test it. It is set to 1 by the library constructor.

The main use-case is to allow applications to detect that they should
not try to close file descriptors that do not belong to them (e.g.  BSD
closefrom). This is a common pattern with applications invoking
daemon(3).

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>