Add support for true globbing with * and ? characters
Currently, the wildcard enabler is limited to match the beginning
of event names.

This patch adds support for true globbing, including the '*' and '?'
globbing characters, which can be inserted anywhere and multiple times
in the globbing expression.

The syntax is a subset of glob(7). The character class and character
range expressions are not implemented as of this patch. '*' matches
zero or more characters, while '?' matches exactly one character.

The '*' globbing character is backward-compatible with the old,
only-at-the-end one. However '?' was a literal character in old
event names, whereas it's a globbing character now. This should not
be a problem in most situations, knowing that LTTng-UST tracepoint
names are limited to C identifiers.

Globbing characters and '\' can be escaped with '\'.

The implementation uses the POSIX basic regular expression (BRE) engine
to avoid duplicating this code. This API is supported by glibc, newlib,
as well as BSD's libc. Globbing characters are converted to their
equivalent BRE expressions, and BRE metacharacters in the globbing
expression are escaped accordingly.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>