jul-agent: Add properties to not log source info
Calculating the source method and class in JUL has a high overhead. If
this information is not necessary, there is benefit in not calculating
it.

This patch allows to configure if we want this information or not. In a
logger.properties file, one can add:

org.lttng.ust.agent.jul.LttngLogHandler.logSourceMethod = false
org.lttng.ust.agent.jul.LttngLogHandler.logSourceClass = false

and the source method/class will not be inferred.

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>