jul-agent: Add property to not log source info
Calculating the source method and class in JUL has a high overhead. If
this information is not necessary, there is benefit in not calculating
it.

This patch allows to configure if we want this information or not. In a
logger.properties file, one can add:

org.lttng.ust.agent.jul.LttngLogHandler.logSource = false

and the source method/class will not be inferred.

Running the lttng-ust-java-tests benchmarks, here are the results for
the 1 and 8 threads loop (results in ns/loop)

* No logger: 9 / 20
* No handler: 112 / 861
* Tracing disabled: 203 / 887
* UST handler with source: 7051 / 13686
* UST handler without source: 1151 / 1980
* File handler: 18533 / 188598

Size of the trace (for the whole benchmark run):

* without source (0.6% lost events): 2.1GB
* with source (0.09% lost events): 5.0GB

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>