Add LTTNG_PACKED ifdefs to validate that it is defined
If LTTNG_PACKED is used to specify whether a structure is packed, but we
end up forgetting inclusion of lttng/ust-compiler.h (which defines it),
we end up silently _not_ packing the data structure, because
LTTNG_PACKED will be considered to be an identifier by the compiler,
and therefore simply ignored.

There are no such instances in lttng-ust, but let's add a ifdef check.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>